refactor: split test_prettytable into multiple files and deined city test data only once
