+++ .pre-commit-config.yaml
@@ -22,7 +22,7 @@ repos:
       - id: end-of-file-fixer
       - id: forbid-submodules
       - id: trailing-whitespace
-        exclude: .github/ISSUE_TEMPLATE/bug_report.md|tests/test_prettytable.py
+        exclude: .github/ISSUE_TEMPLATE/bug_report.md|tests/test_style.py|tests/test_prettytable.py
 
   - repo: https://github.com/python-jsonschema/check-jsonschema
     rev: 0.31.0
