+++ tests/test_colortable.py
@@ -1,37 +1,18 @@
 from __future__ import annotations
 
 import pytest
+from test_prettytable import CITY_DATA, CITY_DATA_HEADER
 
 from prettytable import PrettyTable
 from prettytable.colortable import RESET_CODE, ColorTable, Theme, Themes
 
 
 @pytest.fixture
-def row_prettytable() -> PrettyTable:
-    # Row by row...
-    table = PrettyTable()
-    table.field_names = ["City name", "Area", "Population", "Annual Rainfall"]
-    table.add_row(["Adelaide", 1295, 1158259, 600.5])
-    table.add_row(["Brisbane", 5905, 1857594, 1146.4])
-    table.add_row(["Darwin", 112, 120900, 1714.7])
-    table.add_row(["Hobart", 1357, 205556, 619.5])
-    table.add_row(["Sydney", 2058, 4336374, 1214.8])
-    table.add_row(["Melbourne", 1566, 3806092, 646.9])
-    table.add_row(["Perth", 5386, 1554769, 869.4])
-    return table
-
-
-@pytest.fixture
 def row_colortable():
     table = ColorTable()
-    table.field_names = ["City name", "Area", "Population", "Annual Rainfall"]
-    table.add_row(["Adelaide", 1295, 1158259, 600.5])
-    table.add_row(["Brisbane", 5905, 1857594, 1146.4])
-    table.add_row(["Darwin", 112, 120900, 1714.7])
-    table.add_row(["Hobart", 1357, 205556, 619.5])
-    table.add_row(["Sydney", 2058, 4336374, 1214.8])
-    table.add_row(["Melbourne", 1566, 3806092, 646.9])
-    table.add_row(["Perth", 5386, 1554769, 869.4])
+    table.field_names = CITY_DATA_HEADER
+    for row in CITY_DATA:
+        table.add_row(row)
     return table
